Refactor shared components into dedicated files
